Using which() from the Os utility
